histopathR: Collection Of R Codes Used in Histopathological Research.

Welcome to the histopathR package.

Details

I have been collecting R Codes in another repository: [My R Codes For Data Analysis](https://github.com/sbalci/MyRCodesForDataAnalysis). As I have stated before: > **My aim** is to collect all the codes one needs, where one starts with an excel or spss file and then end with the most common analysis used in histopathology papers : [example table](https://media.nature.com/original/nature-assets/modpathol/journal/v31/n1/extref/modpathol2017106x4.docx) > There are plenty of ways to do an analysis in R, which is great but also confusing for the newbies. I will collect the codes here so that I can refer later and then update them as I learn more.

In this package I would like to combine this effort under a package.

If you would like to help me to collect R codes for histopathological data analysis please contact me, [issue a problem](https://github.com/sbalci/histopathR/issues) or feel free to [make a pull request](https://github.com/sbalci/histopathR/pulls).

histopathR functions

The histopathR functions ...